pyproject.toml
[tool.poetry]
name = "repl_python3_chaatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
chatterbot-corpus = "^1.2.0"
chatterbot-voice = "^0.1.5"
flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
